[REF] split beesdoo_pos
into:
- pos_eater
- pos_available_in_pos_unlink -> cie-pos
- pos_forbid_new_customer -> moved to oca/pos
- pos_partner_no_create -> moved to oca/pos

[FIX] pos_auto_invoice,pos_eater: fix js module names

[FIX] pos_auto_invoice: missing js asset

[FIX] pos_eater: fix description

[FIX] beesdoo_pos: missing dependency

[FIX] pos_eater: data in manifest
