--- beesdoo_pos/data/default_barcode_pattern.xml
@@ -1,7 +1,0 @@
-<odoo>
-    <data>
-        <record id="point_of_sale.barcode_rule_client" model="barcode.rule">
-            <field name="encoding">ean13</field>
-        </record>
-    </data>
-</odoo>
