--- beesdoo_pos/models/__init__.py
@@ -1,2 +1,0 @@
-from . import beesdoo_pos
-from . import product
