+++ beesdoo_pos/readme/DESCRIPTION.rst
@@ -1,5 +1,1 @@
-This module :
-- adds the eaters of the customer to the POS ActionpadWidget and PaymentScreenWidget.
-- maintains "Available in POS" (`available_in_pos`) value when unchecking "Can be Sold" (`sale_ok`).
-- hides phone and address field in pos customer list
-- do not allow customer creation from the POS
+Emptied.
