--- beesdoo_pos/static/src/xml/templates.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates id="template">
-    <t t-extend="ActionpadWidget">
-        <t t-jquery="t[t-if='widget.pos.get_client()']" t-operation="after">
-            <t t-if="widget.pos.get_client()">
-                <div class='customer-information-pay' />
-            </t>
-        </t>
-    </t>
-
-
-    <t t-extend="ClientListScreenWidget">
-        <t t-jquery='.new-customer' t-operation='replace'>
-        </t>
-
-        <t t-jquery=".client-list" t-operation="replace">
-
-            <table class='client-list'>
-                <thead>
-                    <tr>
-                        <th>Name</th>
-                    </tr>
-                </thead>
-                <tbody class='client-list-contents'>
-                </tbody>
-            </table>
-
-        </t>
-    </t>
-
-    <t t-extend="ClientLine">
-        <t t-jquery=".client-line" t-operation="replace">
-            <tr class='client-line' t-att-data-id='partner.id'>
-                <td>
-                    <t t-esc='partner.name' />
-                </td>
-            </tr>
-        </t>
-    </t>
-
-    <t t-extend="ClientDetails">
-        <t t-jquery=".client-details-left" t-operation="replace">
-            <div class='client-details-left' />
-        </t>
-        <t t-jquery=".edit-buttons" t-operation="replace">
-            <div class='edit-buttons' />
-        </t>
-    </t>
-</templates>
