--- beesdoo_pos/views/beesdoo_pos.xml
@@ -1,11 +1,0 @@
-<odoo>
-    <template id="assets" inherit_id="point_of_sale.assets">
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/beesdoo_pos/static/src/js/beesdoo.js"
-            />
-            <link rel='stylesheet' href="/beesdoo_pos/static/src/css/beesdoo.css" />
-        </xpath>
-    </template>
-</odoo>
