+++ pos_auto_invoice_company/i18n/pos_auto_invoice_company.pot
@@ -1,0 +1,32 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* beesdoo_pos
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: beesdoo_pos
+#: model:ir.model,name:beesdoo_pos.model_res_partner
+msgid "Contact"
+msgstr ""
+
+#. module: beesdoo_pos
+#. openerp-web
+#: code:addons/beesdoo_pos/static/src/xml/templates.xml:21
+#, python-format
+msgid "Name"
+msgstr ""
+
+#. module: beesdoo_pos
+#: model:ir.model,name:beesdoo_pos.model_product_template
+msgid "Product Template"
+msgstr ""
+
