+++ pos_auto_invoice_company/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* BEES coop - Cellule IT
+* Coop IT Easy SC
