+++ pos_auto_invoice_company/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Applies to_invoice to company partners.
