+++ pos_auto_invoice_company/static/src/css/pos_auto_invoice_company.css
