+++ pos_eater/README.rst
@@ -1,0 +1,63 @@
+=======================
+BEES coop Point of sale
+=======================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-beescoop%2FObeesdoo-lightgray.png?logo=github
+    :target: https://github.com/beescoop/Obeesdoo/tree/12.0/beesdoo_pos
+    :alt: beescoop/Obeesdoo
+
+|badge1| |badge2| |badge3| 
+
+This module :
+- adds the eaters of the customer to the POS ActionpadWidget and PaymentScreenWidget.
+- maintains "Available in POS" (`available_in_pos`) value when unchecking "Can be Sold" (`sale_ok`).
+- hides phone and address field in pos customer list
+- do not allow customer creation from the POS
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/beescoop/Obeesdoo/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/beescoop/Obeesdoo/issues/new?body=module:%20beesdoo_pos%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* BEES coop - Cellule IT
+* Coop IT Easy SC
+
+Contributors
+~~~~~~~~~~~~
+
+* BEES coop - Cellule IT
+* Coop IT Easy SC
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `beescoop/Obeesdoo <https://github.com/beescoop/Obeesdoo/tree/12.0/beesdoo_pos>`_ project on GitHub.
+
+You are welcome to contribute.
