+++ pos_eater/__manifest__.py
@@ -8,19 +8,16 @@
 #   - Houssine Bakkali <houssine@coopiteasy.be>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {
-    "name": "BEES coop Point of sale",
-    "summary": """Emptied.""",
+    "name": "POS Eater",
+    "summary": """This module adds the eaters of the customer to the POS
+    ActionpadWidget.""",
     "author": "BEES coop - Cellule IT, Coop IT Easy SC",
     "website": "https://github.com/beescoop/Obeesdoo",
     "category": "Point Of Sale",
-    "version": "12.0.2.0.0",
-    "depends": [
-        "pos_eater",
-        "pos_available_in_pos_unlink",
-        "pos_partner_no_create",
-        "pos_hide_partner_info",
-        "pos_auto_invoice_company",
-    ],
+    "version": "12.0.1.0.0",
+    "depends": ["point_of_sale", "eater"],
+    "data": ["views/assets.xml"],
+    "qweb": ["static/src/xml/templates.xml"],
     "installable": True,
     "license": "AGPL-3",
 }
