+++ pos_eater/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* BEES coop - Cellule IT
+* Coop IT Easy SC
