+++ pos_eater/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Adds the eaters of the customer to the POS ActionpadWidget.
