+++ pos_eater/views/assets.xml
@@ -1,0 +1,7 @@
+<odoo>
+    <template id="assets" inherit_id="point_of_sale.assets">
+        <xpath expr=".">
+            <script type="text/javascript" src="/pos_eater/static/src/js/screens.js" />
+        </xpath>
+    </template>
+</odoo>
