+++ setup/pos_auto_invoice_company/odoo/addons/pos_auto_invoice_company
@@ -1,0 +1,1 @@
+../../../../pos_auto_invoice_company
