+++ setup/pos_eater/odoo/addons/pos_eater
@@ -1,0 +1,1 @@
+../../../../pos_eater
