+++ test-requirements.txt
@@ -1,7 +1,9 @@
-git+https://github.com/coopiteasy/vertical-cooperative@12.0#subdirectory=setup/portal_recaptcha
 git+https://github.com/coopiteasy/addons@12.0#subdirectory=setup/email_template_config
 git+https://github.com/coopiteasy/addons@12.0#subdirectory=setup/partner_contact_address
 git+https://github.com/coopiteasy/addons@12.0#subdirectory=setup/company_supplier_context
+git+https://github.com/coopiteasy/cie-pos@12.0#subdirectory=setup/pos_available_in_pos_unlink
+git+https://github.com/coopiteasy/vertical-cooperative@12.0#subdirectory=setup/portal_recaptcha
 git+https://github.com/grap/grap-odoo-incubator@12.0#subdirectory=setup/product_print_category
+
 git+https://github.com/coopiteasy/account-invoicing@12.0-account_invoice_date_required#subdirectory=setup/account_invoice_date_required
 git+https://github.com/coopiteasy/account-invoicing@12.0-account_invoice_negative_total#subdirectory=setup/account_invoice_negative_total
